refactor code of main
