+++ API/Routers/faculty.py
@@ -1,0 +1,3 @@
+"""
+APP Faculty for handling Lectures, Subjects and hours for each subject
+"""
