+++ API/Routers/planer.py
@@ -1,0 +1,3 @@
+"""
+APP Planer for planing hours for Lectures in Faculty
+"""
