+++ API/Routers/schedule.py
@@ -1,0 +1,3 @@
+"""
+APP Schedule for handling Schedules for each subject and lectures
+"""
