+++ pyproject.toml
@@ -7,17 +7,50 @@ readme = "README.md"
 packages = [{include = "fh_gateway"}]
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "~3.11"
 fastapi = "^0.109.2"
 uvicorn = "^0.27.1"
-PyJWT = "^2.8.0"
-python-dotenv = "^1.0.1"
-pre-commit = "^3.6.2"
-authlib = "^1.3.0"
+PyJWT = "~2.8"
+python-dotenv = "~1.0"
+authlib = "~1.3"
 httpx = "^0.27.0"
 itsdangerous = "^2.1.2"
 
+[tool.poetry.group.dev]
+optional = true
+
+[tool.poetry.group.dev.dependencies]
+black = "*"
+isort = "*"
+pre-commit = "~3.6"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.black]
+line-length = 120
+exclude = '''
+(
+    /(
+        \.git
+      | \.hg
+      | \.mypy_cache
+      | \.tox
+      | \.env
+      | _build
+      | buck-out
+      | build
+      | dist
+    )/
+  | .*/migrations
+  | .*/settings
+  | .*/__init__.py
+  | .*/static
+  | .*/media
+)
+'''
+
+[tool.isort]
+profile = "black"
+skip = ["**/migrations", "**/settings", "**/__init__.py", "**/static", "**/media"]
