--- templates/index.html
@@ -1,1 +1,0 @@
-Main site - index
