--- templates/profile.html
@@ -1,1 +1,0 @@
-{{ userinfo }}
